clash/requirements.txt
python-dotenv==0.21.0
requests
pandas
numpy